# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Shadow Garden API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/user/stats:
    get:
      operationId: getUserStats
      tags: [user]
      summary: Get current user full profile and stats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/user/inventory:
    get:
      operationId: getUserInventory
      tags: [user]
      summary: Get current user inventory
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/user/achievements:
    get:
      operationId: getUserAchievements
      tags: [user]
      summary: Get current user achievements and badges
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Achievements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/guilds:
    get:
      operationId: getGuilds
      tags: [guilds]
      summary: Get all guilds
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Guilds list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildsResponse"
  /v1/guilds/{guildId}:
    get:
      operationId: getGuild
      tags: [guilds]
      summary: Get a specific guild by ID
      parameters:
        - in: path
          name: guildId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Guild details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuildDetails"
        "404":
          description: Guild not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        level:
          type: integer
        xp:
          type: integer
        balance:
          type: integer
        bank:
          type: integer
        premium:
          type: integer
        bio:
          type: string
        registeredAt:
          type: integer
      required:
        - id
        - name
        - level
        - xp
        - balance
        - bank
    UserStats:
      type: object
      properties:
        profile:
          $ref: "#/components/schemas/UserProfile"
        rpg:
          $ref: "#/components/schemas/RpgStats"
        guild:
          $ref: "#/components/schemas/GuildMembership"
        rank:
          type: integer
        totalUsers:
          type: integer
        xpNeeded:
          type: integer
      required:
        - profile
        - rank
        - totalUsers
        - xpNeeded
    RpgStats:
      type: object
      properties:
        class:
          type: string
        hp:
          type: integer
        maxHp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        speed:
          type: integer
        dungeonFloor:
          type: integer
      required:
        - class
        - hp
        - maxHp
        - attack
        - defense
        - speed
        - dungeonFloor
    GuildMembership:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        level:
          type: integer
      required:
        - id
        - name
        - level
    InventoryItem:
      type: object
      properties:
        item:
          type: string
        quantity:
          type: integer
        category:
          type: string
      required:
        - item
        - quantity
        - category
    InventoryResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
      required:
        - items
    Achievement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        earnedAt:
          type: integer
      required:
        - id
        - name
        - description
        - icon
        - earnedAt
    AchievementsResponse:
      type: object
      properties:
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
      required:
        - achievements
...
    Guild:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        level:
          type: integer
        memberCount:
          type: integer
        ownerName:
          type: string
        createdAt:
          type: integer
      required:
        - id
        - name
        - level
        - memberCount
        - ownerName
    GuildsResponse:
      type: object
      properties:
        guilds:
          type: array
          items:
            $ref: "#/components/schemas/Guild"
      required:
        - guilds
    GuildDetails:
      type: object
      properties:
        guild:
          $ref: "#/components/schemas/Guild"
        members:
          type: array
          items:
            type: object
            properties:
              userId:
                type: string
              name:
                type: string
              level:
                type: integer
              joinedAt:
                type: integer
            required:
              - userId
              - name
              - level
              - joinedAt
      required:
        - guild
        - members
...